[INLONG-10558][Manager] Rename installerCode
