+++ inlong-agent/agent-installer/src/main/java/org/apache/inlong/agent/installer/ManagerFetcher.java
@@ -25,7 +25,7 @@ import org.apache.inlong.agent.utils.HttpManager;
 import org.apache.inlong.agent.utils.ThreadUtils;
 import org.apache.inlong.common.pojo.agent.installer.ConfigRequest;
 import org.apache.inlong.common.pojo.agent.installer.ConfigResult;
-import org.apache.inlong.common.pojo.agent.installer.InstallerCode;
+import org.apache.inlong.common.pojo.agent.AgentResponseCode;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -124,7 +124,7 @@ public class ManagerFetcher extends AbstractDaemon implements ProfileFetcher {
             while (isRunnable()) {
                 try {
                     ConfigResult config = getConfig();
-                    if (config != null && config.getCode().equals(InstallerCode.SUCCESS)) {
+                    if (config != null && config.getCode().equals(AgentResponseCode.SUCCESS)) {
                         manager.getModuleManager().submitConfig(config);
                     }
                 } catch (Throwable ex) {
