+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/agent/AgentConfigInfo.java
@@ -17,8 +17,6 @@
 
 package org.apache.inlong.common.pojo.agent;
 
-import org.apache.inlong.common.pojo.agent.installer.InstallerCode;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -33,7 +31,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AgentConfigInfo {
 
-    InstallerCode code;
+    AgentResponseCode code;
     private String zkUrl;
     private AgentClusterInfo cluster;
     private String md5;
