+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/agent/TaskResult.java
@@ -17,8 +17,6 @@
 
 package org.apache.inlong.common.pojo.agent;
 
-import org.apache.inlong.common.pojo.agent.installer.InstallerCode;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -38,6 +36,6 @@ public class TaskResult {
     private List<CmdConfig> cmdConfigs;
     private List<DataConfig> dataConfigs;
     private String md5;
-    InstallerCode code;
+    AgentResponseCode code;
 
 }
