+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/agent/installer/ConfigResult.java
@@ -21,6 +21,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.apache.inlong.common.pojo.agent.AgentResponseCode;
 
 import java.util.List;
 
@@ -36,7 +37,7 @@ public class ConfigResult {
     /**
      * The code of the config result
      */
-    InstallerCode code;
+    AgentResponseCode code;
 
     /**
      * The md5 of the config result
